requirements.txt
matplotlib==2.2.2
Keras==2.2.2
numpy==1.14.3
opencv_python==3.4.0.12
pandas==0.23.4
tensorflow==1.10.1
Pillow==5.2.0
scikit_learn==0.19.2
